pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "cvgen"
version = "1.0.0"
description = "Quantum Computing for Every Device — AI Agents, Visual Circuit Builder, Cloud Quantum Hardware"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
authors = [
    { name = "aigambitkg" },
]
keywords = ["quantum", "ai-agents", "quantum-computing", "origin-pilot", "quantum-api", "quantum-web", "ibm-quantum", "aws-braket", "azure-quantum"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Scientific/Engineering :: Physics",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "numpy>=1.24",
    "scipy>=1.10",
    "pyzmq>=25.0",
]

[project.optional-dependencies]
zmq = ["pyzmq>=25.0"]
rag = ["qdrant-client>=1.7", "httpx>=0.27"]
llm = ["httpx>=0.27"]
api = [
    "fastapi>=0.110",
    "uvicorn[standard]>=0.27",
    "websockets>=12.0",
]
qiskit = ["qiskit>=1.0"]
ibm = ["qiskit>=1.0", "qiskit-ibm-runtime>=0.20"]
braket = ["amazon-braket-sdk>=1.70"]
azure = ["azure-quantum>=1.0", "qiskit>=1.0"]
origin = ["pyqpanda3"]
all-backends = [
    "qiskit>=1.0",
    "qiskit-ibm-runtime>=0.20",
    "amazon-braket-sdk>=1.70",
    "azure-quantum>=1.0",
    "pyqpanda3",
]
full = [
    "pyzmq>=25.0",
    "qdrant-client>=1.7",
    "httpx>=0.27",
    "fastapi>=0.110",
    "uvicorn[standard]>=0.27",
    "websockets>=12.0",
    "qiskit>=1.0",
    "qiskit-ibm-runtime>=0.20",
    "amazon-braket-sdk>=1.70",
    "azure-quantum>=1.0",
    "pyqpanda3",
]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "httpx>=0.27",
    "ruff>=0.4",
    "pyzmq>=25.0",
    "mypy>=1.0",
]

[project.scripts]
cvgen = "cvgen.cli:main"

[project.urls]
Documentation = "https://github.com/aigambitkg/CVGen/blob/main/docs"
Repository = "https://github.com/aigambitkg/CVGen"
Issues = "https://github.com/aigambitkg/CVGen/issues"
Changelog = "https://github.com/aigambitkg/CVGen/blob/main/CHANGELOG.md"

[tool.hatch.build.targets.wheel]
packages = ["src/cvgen"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint.per-file-ignores]
"src/cvgen/core/gates.py" = ["E741"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
ignore_missing_imports = true
